AA-712: fixing up PR comments
